+++ app/src/main/java/com/tolstykh/eatABurrita/widget/BurritoWidgetReceiver.kt
@@ -1,0 +1,8 @@
+package com.tolstykh.eatABurrita.widget
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+
+class BurritoWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = BurritoWidget()
+}